+++ package.json
@@ -22,10 +22,10 @@
         "test": "npm run cleantsc && jest",
         "tsc": "tsc",
         "clean": "rm -rf build/",
-        "cleantsc": "rm -f index.js *.d.ts handler/*.js handler/*.d.ts",
+        "cleantsc": "rm -f index.js cdkv1.js cdkv2.js *.d.ts handler/*.js handler/*.d.ts",
         "compile": "npm run clean && npm run -- tsc --outDir build/ && cp package.json README.md ./build/",
         "version": "auto-changelog --hide-credit -p && git add CHANGELOG.md",
-        "_lint": "eslint 'index.ts' 'handler/**/*.ts' 'test/**/*.ts' jest.config.js",
+        "_lint": "eslint 'index.ts' cdkv1.ts cdkv2.ts 'handler/**/*.ts' 'test/**/*.ts' jest.config.js",
         "lint": "npm run -- _lint --fix",
         "lint-check": "npm run -- _lint --fix --max-warnings 0",
         "docs": "typedoc index.ts --includeVersion"
@@ -44,8 +44,10 @@
         "@typescript-eslint/eslint-plugin": "^5.0.0",
         "@typescript-eslint/parser": "^5.0.0",
         "auto-changelog": "^2.3.0",
+        "aws-cdk-lib": "^2.15.0",
         "aws-sdk": "^2.1025.0",
         "aws-sdk-mock": "^5.4.0",
+        "constructs": "^10.0.0",
         "eslint": "^8.0.1",
         "eslint-config-prettier": "^8.3.0",
         "eslint-plugin-prettier": "^4.0.0",
