+++ tsconfig.json
@@ -22,6 +22,8 @@
   },
   "include": [
     "index.ts",
+    "cdkv1.ts",
+    "cdkv2.ts",
     "handler/"
   ],
   "exclude": [
